Extract cardboard/api HTTP transport to app/utils
